восстановил освещение, события кликов/дижений мыщи

+++ OpenGLReborn/MouseMoveEvent.cpp
@@ -1,6 +1,6 @@
 #include "MouseMoveEvent.h"
 
-MouseMoveEvent::MouseMoveEvent(float _xPosition, float _yPosition, float _xOffset, float _yOffset):
+MouseMoveEvent::MouseMoveEvent(double _xPosition, double _yPosition, double _xOffset, double _yOffset):
 	UIEvent(MOUSE_MOVE_EVENT, _xPosition, _yPosition),
 	xOffset(_xOffset),
 	yOffset(_yOffset)
